requirements.txt
   streamlit
   python-docx
   pdfplumber
   requests